[DCK] Increase SHM size for Postgres
This was added to the official image docs in https://github.com/docker-library/docs/pull/1331 and Odoo is heavy on Postgres, so we add it as a default value.

This will prevent unexpected failures on big transactions, such as migrations. The recommended value of 256mb proved to be still not enough for some cases, so I'm duplicating it here.