Add a docker-compose yaml file for remote development
Developing on a remote system requires an encrypted reverse proxy.
This file is just a copy of the normal devel file, but with
additions for using a Let's Encryp enabled Traefik reverse-proxy.
To use it:
in the .env file modify DOMAIN_PROD with the remote machine host name
docker-compose -f remote-devel.yaml up
docker-compose -p inverseproxy -f inverseproxy.yaml up

TODO: refactor the file to depend on devel.yaml instead